Provide main function FitModel and SagaSolver
* FitModel provides the link to R. Its
only real purpose is to convert the
feature matrix to arma::mat or arma::sp_mat
and pass it along to SagaSolver.
* SagaSolver is templated to support both
sparse and dense feature matrices and
houses all the fitting procedure. In the
future, this should likely be split up
into smaller parts to better support
fitting paths of lambda values.