Add a plotting method for sgdnet objects
This plotting method mimics plot.glmnet() but uses lattice graphics instead of base graphics to make it easier to deal with the output since it is now a grid/trellis object and makes it much easier to modify the arguments of the underlying function since we are calling a single function behind the scenes (xyplot()).

We also add a simple test for the gaussian regression to check that we retrieve a trellis object. We will also need to use better plot tests where the actual plotting mechanism is called eventually, but for now I think this will do.