Refresh and improve documentation
Change scaling to be consistent with glmnet

Rearrange and drop scaling in glm() comparison since result should be equivalent

Explicitly set seed in gaussian regression test

Switch from armadillo to STL containers

This update refactors the internals of sgdnet to use STL-type vectors instead of their armadillo counterparts. The main reason for this is subpar performance. We seldom use the specialized functionality of the armadillo objects and so can benefit from the leaner STL objects. Here is a brief description of the changes:

* We now use STL vectors instead of matrices for the response, weights, and gradient memory (and more), which means that we now index inside the vector by, for instance for the weights, feature_index*number_of_classes + class+index, which means we don't have to use a matrix for univariate responses.
* Loss() and Gradient() in the Family class now instead returns and works on doubles instead of row vectors.
* I have written separate Mean and StandardDeviation functions now that we are not using the armadillo interface as much.

This commit also addresses a long-standing issue with the elastic net penalty that was not consistent with the results from glmnet() but now is.

Switch from armadillo to STL containers

This update refactors the internals of sgdnet to use STL-type vectors instead of their armadillo counterparts. The main reason for this is subpar performance. We seldom use the specialized functionality of the armadillo objects and so can benefit from the leaner STL objects. Here is a brief description of the changes:

* We now use STL vectors instead of matrices for the response, weights, and gradient memory (and more), which means that we now index inside the vector by, for instance for the weights, feature_index*number_of_classes + class+index, which means we don't have to use a matrix for univariate responses.
* Loss() and Gradient() in the Family class now instead returns and works on doubles instead of row vectors.
* I have written separate Mean and StandardDeviation functions now that we are not using the armadillo interface as much.

This commit also addresses a long-standing issue with the elastic net penalty that was not consistent with the results from glmnet() but now is.

Modify step size to match previous definition of alpha penalty

Normalize loss

Make results comparable with binomial branch

Fix erroneous timing of rescaling

Refresh and improve documentation

Change scaling to be consistent with glmnet

Normalize loss

Make results comparable with binomial branch

Fix erroneous timing of rescaling

Refresh and improve documentation

Change scaling to be consistent with glmnet